move roms fixtures into roms/conftest.py, define fixtures for fakes used in fake_romssimulation, new fixture returning example simulation as a dict, there are now single definition points for all faked values
